Fix scale line units change depending on zoom.
OpenLayers ScaleLine has topInUnit, topOutUnit, bottomInUnit, and
bottomOutUnit parameters.  We were only setting the *OutUnit
parameters before.  This sets the *InUnit parameters to match
the *OutUnit parameters.

Note: Sometimes the ScaleLine appears to not be honoring the
maxWidth parameter.  This appears to be an OpenLayers bug, not
sure what we can do about it.

refs: geomoose/geomoose#102 (FYI from tchaddad in issue comment).